command_line/chat/requirements.txt
Flask
google-generativeai
python-dotenv # Good practice, though we're using json for this specific request